[feat] Open API 데이터베이스 저장 로직 재정립 및 엔드포인트 MovieSearchList By 검색어 개발.

+++ kickmovie/src/main/java/kh/gangnam/kickmovie/dto/MovieDetailDTO.java
@@ -17,11 +17,11 @@ import java.util.List;
 public class MovieDetailDTO {
 
     @JsonProperty("id")
-    private Long movieId;
+    private long movieId;
     private boolean adult;
     @JsonProperty("belongs_to_collection")
     private BelongsToCollectionDTO belongsToCollection;
-    private int budget;
+    private long budget;
     private List<GenreDTO> genres;
     @JsonProperty("homepage")
     private String homePage;
@@ -41,8 +41,8 @@ public class MovieDetailDTO {
     private List<ProductionCountryDTO> productionCountries;
     @JsonProperty("release_date")
     private String releaseDate;
-    private int revenue;
-    private int runtime;
+    private long revenue;
+    private long runtime;
     @JsonProperty("spoken_languages")
     private List<SpokenLanguageDTO> spokenLanguages;
     private String status;
@@ -52,5 +52,5 @@ public class MovieDetailDTO {
     @JsonProperty("vote_average")
     private Double voteAverage;
     @JsonProperty("vote_count")
-    private int voteCount;
+    private long voteCount;
 }
